fix ptr

+++ jacobi.h
@@ -2,4 +2,4 @@
 
 void jacobistep(float *psinew, float *psi, int m, int n);
 
-double deltasq(double *newarr, double *oldarr, int m, int n);
+double deltasq(float *newarr, float *oldarr, int m, int n);
